feat: endpoints

+++ src/main/java/com/springapitest/restapi/repositories/TransactionRepository.java
@@ -2,6 +2,7 @@ package com.springapitest.restapi.repositories;
 import org.springframework.data.repository.CrudRepository;
 import com.springapitest.restapi.models.Transaction;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
+
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
 
 }
